Separate forms in steps (#6108)
* Survey steps

* Add basic system tests

* Add separators in admin area

* Persist separators

* Fix weird behavior when adding questions and moving them up

they were being added to the end of the element, but there are some invisible elements there, so moving the question up seemed not to work for a while, but it was actually moving betwenn invisible elements. This is now fixed, since new questions are added right after the last one, thus avoiding the invisible elements.

* Fix bug when adding a separator, buttons where not correct

* Fix bug when adding questions on new survey

* Hide separator from question types list

* Fix bug on events binding

This made the binding to fire for every click on the wrapper element,
thus creating a lot of child answer options.

* Prepare for step layout in public area

* Paginate steps in public area

* Handle survey submissions

* Delete unused locale keys

* Fix tests for public area on surveys

* Generalize use case

* Add missing tests

* Add changelog

* Fix wrong locale key

* Hide spearators for anonymous users

* Simplify code

* Fix refactored method

* Surveys font style fixes

Co-authored-by: Javier Usobiaga <javier@swwweet.com>
# Conflicts:
#	decidim_app-design/app/views/public/index.html.erb